+++ src/debug.h
@@ -1,6 +1,8 @@
 #ifndef _DEBUG_H_
 #define _DEBUG_H_
 
+#include <array>
+
 #include "conddebug.h"
 #include "git.h"
 #include "nsf.h"
@@ -120,7 +122,9 @@ extern void IncrementInstructionsCounters();
 
 //internal variables that debuggers will want access to
 extern uint8 *vnapage[4],*VPage[8];
-extern uint8 PPU[4],PALRAM[0x20],UPALRAM[3],SPRAM[0x100],VRAMBuffer,PPUGenLatch,XOffset;
+extern uint8 PPU[4],SPRAM[0x100],VRAMBuffer,PPUGenLatch,XOffset;
+extern std::array<uint8_t, 0x20> PALRAM;
+extern std::array<uint8_t, 3> UPALRAM;
 extern uint32 FCEUPPU_PeekAddress();
 extern uint8 READPAL_MOTHEROFALL(uint32 A);
 extern int numWPs;